Feature: players now get score for gun kills, chat messages buffered (but do not show if a player has disconnected).

+++ server/src/main/java/com/limpygnome/projectsandbox/server/players/SessionMetrics.java
@@ -40,8 +40,9 @@ public class SessionMetrics
         this.dirty = true;
     }
 
-    public synchronized void incrementScore(long amount)
+    public synchronized void incrementScore(int amount)
     {
+        // TODO: check this works for negatives and high longs
         this.score += amount;
         this.dirty = true;
     }
